class lintul_cassava_meteo renamed to lintul_cassava_penman.
